@@ -44,7 +44,7 @@
     "@typechain/ethers-v5": "^9.0.0",
     "@typechain/hardhat": "^5.0.0",
     "@types/node": "^16.11.7",
-    "@types/react": "^16.9.50",
+    "@types/react": "^18.0.0",
     "@types/react-dom": "^16.9.5",
     "@uiw/react-md-editor": "1.14.0",
     "@uniswap/token-lists": "^1.0.0-beta.27",
@@ -81,12 +81,12 @@
     "polished": "^3.4.4",
     "promievent": "^0.1.3",
     "qrcode.react": "^1.0.0",
-    "react": "^16.10.2",
+    "react": "^18.0.0",
     "react-app-rewired": "^2.2.1",
     "react-countdown": "^2.3.2",
     "react-datetime": "^3.1.1",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^16.10.2",
+    "react-dom": "^18.0.0",
     "react-google-charts": "^3.0.15",
     "react-icons": "^4.3.1",
     "react-intersection-observer": "^8.33.1",
