@@ -1,8 +1,11 @@
 import { EtherSWRConfig } from 'ether-swr';
 import useJsonRpcProvider from 'hooks/Guilds/web3/useJsonRpcProvider';
 import loggerMiddleware from 'hooks/Guilds/ether-swr/middleware/logger';
+import React from 'react';
 
-const EtherSWRManager: React.FC = ({ children }) => {
+const EtherSWRManager: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
   const provider = useJsonRpcProvider();
 
   return (
